Fix codestyle a bit
