+++ src/main/java/ok/dht/test/kosnitskiy/ServiceImpl.java
@@ -7,10 +7,10 @@ import jdk.incubator.foreign.MemorySegment;
 import ok.dht.Service;
 import ok.dht.ServiceConfig;
 import ok.dht.test.ServiceFactory;
-import ok.dht.test.kosnitskiy.DAO.BaseEntry;
-import ok.dht.test.kosnitskiy.DAO.Config;
-import ok.dht.test.kosnitskiy.DAO.Entry;
-import ok.dht.test.kosnitskiy.DAO.MemorySegmentDao;
+import ok.dht.test.kosnitskiy.dao.BaseEntry;
+import ok.dht.test.kosnitskiy.dao.Config;
+import ok.dht.test.kosnitskiy.dao.Entry;
+import ok.dht.test.kosnitskiy.dao.MemorySegmentDao;
 import ok.dht.test.kosnitskiy.server.HttpServerImpl;
 import one.nio.http.HttpServerConfig;
 import one.nio.http.Param;
