+++ src/main/java/ok/dht/test/kosnitskiy/dao/MemorySegmentComparator.java
@@ -1,11 +1,11 @@
-package ok.dht.test.kosnitskiy.DAO;
+package ok.dht.test.kosnitskiy.dao;
 
 import jdk.incubator.foreign.MemoryAccess;
 import jdk.incubator.foreign.MemorySegment;
 
 import java.util.Comparator;
 
-public class MemorySegmentComparator implements Comparator<MemorySegment> {
+public final class MemorySegmentComparator implements Comparator<MemorySegment> {
 
     public static final Comparator<MemorySegment> INSTANCE = new MemorySegmentComparator();
