+++ src/main/java/ok/dht/test/kosnitskiy/dao/PeekIterator.java
@@ -1,4 +1,4 @@
-package ok.dht.test.kosnitskiy.DAO;
+package ok.dht.test.kosnitskiy.dao;
 
 import java.util.Iterator;
