+++ src/main/java/ok/dht/test/kosnitskiy/dao/StorageClosedException.java
@@ -1,4 +1,4 @@
-package ok.dht.test.kosnitskiy.DAO;
+package ok.dht.test.kosnitskiy.dao;
 
 public class StorageClosedException extends RuntimeException {
